+++ extensions/kafka-connect/src/test/java/com/hazelcast/jet/kafka/connect/impl/ConnectorWrapperTest.java
@@ -40,7 +40,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 public class ConnectorWrapperTest {
     @Test
     public void should_create_and_start_source_with_minimal_properties() {
-        new ConnectorWrapper(0L, minimalProperties());
+        new ConnectorWrapper(minimalProperties());
 
         assertThat(sourceConnectorInstance().isInitialized()).isTrue();
         assertThat(sourceConnectorInstance().isStarted()).isTrue();
@@ -48,7 +48,7 @@ public class ConnectorWrapperTest {
 
     @Test
     public void should_create_task_runners() {
-        ConnectorWrapper connectorWrapper = new ConnectorWrapper(0L, minimalProperties());
+        ConnectorWrapper connectorWrapper = new ConnectorWrapper(minimalProperties());
 
         TaskRunner taskRunner1 = connectorWrapper.createTaskRunner(0);
         assertThat(taskRunner1.name()).isEqualTo("some-name-task-0");
@@ -72,7 +72,7 @@ public class ConnectorWrapperTest {
     @Ignore
     @Test
     public void should_reconfigure_task_runners() {
-        ConnectorWrapper connectorWrapper = new ConnectorWrapper(0L, minimalProperties());
+        ConnectorWrapper connectorWrapper = new ConnectorWrapper(minimalProperties());
 
         TaskRunner taskRunner1 = connectorWrapper.createTaskRunner(0);
         assertThat(taskRunner1.name()).isEqualTo("some-name-task-0");
@@ -108,7 +108,7 @@ public class ConnectorWrapperTest {
         Properties properties = new Properties();
         properties.setProperty("name", "some-name");
         properties.setProperty("connector.class", "com.example.non.existing.Connector");
-        assertThatThrownBy(() -> new ConnectorWrapper(0L, properties))
+        assertThatThrownBy(() -> new ConnectorWrapper(properties))
                 .isInstanceOf(HazelcastException.class)
                 .hasMessage("Connector class 'com.example.non.existing.Connector' not found. " +
                         "Did you add the connector jar to the job?");
@@ -118,7 +118,7 @@ public class ConnectorWrapperTest {
     public void should_cleanup_on_destroy() {
         Properties properties = minimalProperties();
         properties.setProperty(ITEMS_SIZE, String.valueOf(3));
-        ConnectorWrapper connectorWrapper = new ConnectorWrapper(0L, properties);
+        ConnectorWrapper connectorWrapper = new ConnectorWrapper(properties);
         assertThat(sourceConnectorInstance().isStarted()).isTrue();
 
         connectorWrapper.stop();
